+++ jest.config.js
@@ -6,10 +6,10 @@ module.exports = {
   moduleDirectories: ["node_modules", "src"],
   moduleFileExtensions: ["ts", "tsx", "js", "jsx"],
   transform: {
-    "node_modules/(map-obj|camelcase)/.+\\.(j|t)sx?$": "ts-jest",
+    "node_modules/(map-obj|camelcase|plur)/.+\\.(j|t)sx?$": "ts-jest",
   },
   testPathIgnorePatterns: ["/node_modules/"],
   testRegex: ".test.(tsx?)$",
 
-  transformIgnorePatterns: [`/node_modules/(?!(map-obj|camelcase))`],
+  transformIgnorePatterns: [`/node_modules/(?!(map-obj|camelcase|plur))`],
 };