+++ src/lib/pages/pools/data.ts
@@ -26,7 +26,9 @@ export const usePools = (
   offset: number,
   pageSize: number
 ): { pools: Option<Pool[]>; isLoading: boolean } => {
-  const { assetInfos, isLoading: isLoadingAssetInfos } = useAssetInfos();
+  const { assetInfos, isLoading: isLoadingAssetInfos } = useAssetInfos({
+    withPrices: true,
+  });
   const { data: poolList, isLoading: isLoadingPoolList } = usePoolListQuery({
     isSupported,
     poolType,
@@ -56,7 +58,9 @@ export const usePools = (
 export const usePool = (
   poolId: number
 ): { pool: Option<PoolDetail>; isLoading: boolean } => {
-  const { assetInfos, isLoading: isLoadingAssetInfos } = useAssetInfos();
+  const { assetInfos, isLoading: isLoadingAssetInfos } = useAssetInfos({
+    withPrices: true,
+  });
   const { data: pool, isLoading: isLoadingPoolInfo } = usePoolByPoolId(poolId);
 
   if (!Number.isInteger(poolId) || poolId <= 0)