+++ src/lib/services/proposalService.ts
@@ -3,7 +3,11 @@ import { useQuery } from "@tanstack/react-query";
 import big from "big.js";
 import { useCallback } from "react";
 
-import { useBaseApiRoute, useCelatoneApp } from "lib/app-provider";
+import {
+  CELATONE_QUERY_KEYS,
+  useBaseApiRoute,
+  useCelatoneApp,
+} from "lib/app-provider";
 import {
   getProposalList,
   getProposalListCount,
@@ -77,7 +81,7 @@ export const useRelatedProposalsByContractAddressPagination = (
 
   return useQuery(
     [
-      "related_proposals_by_contract_address_pagination",
+      CELATONE_QUERY_KEYS.RELATED_PROPOSALS_BY_CONTRACT_ADDRESS_PAGINATION,
       contractAddress,
       offset,
       pageSize,
@@ -109,7 +113,7 @@ export const useRelatedProposalsCountByContractAddress = (
 
   return useQuery(
     [
-      "related_proposals_count_by_contract_address",
+      CELATONE_QUERY_KEYS.RELATED_PROPOSALS_BY_CONTRACT_ADDRESS_COUNT,
       contractAddress,
       indexerGraphClient,
     ],
@@ -151,7 +155,7 @@ export const useProposalsByWalletAddressPagination = (
 
   return useQuery(
     [
-      "proposals_by_wallet_address_pagination",
+      CELATONE_QUERY_KEYS.PROPOSALS_BY_WALLET_ADDRESS_PAGINATION,
       walletAddress,
       indexerGraphClient,
       offset,
@@ -179,7 +183,11 @@ export const useProposalsCountByWalletAddress = (
   }, [walletAddress, indexerGraphClient]);
 
   return useQuery(
-    ["proposals_count_by_wallet_address", walletAddress, indexerGraphClient],
+    [
+      CELATONE_QUERY_KEYS.PROPOSALS_BY_WALLET_ADDRESS_COUNT,
+      walletAddress,
+      indexerGraphClient,
+    ],
     queryFn,
     {
       keepPreviousData: true,
@@ -228,7 +236,13 @@ export const useProposalList = (
     [indexerGraphClient, offset, pageSize, expression]
   );
   return useQuery(
-    ["proposal_list", indexerGraphClient, expression, offset, pageSize],
+    [
+      CELATONE_QUERY_KEYS.PROPOSALS,
+      indexerGraphClient,
+      expression,
+      offset,
+      pageSize,
+    ],
     queryFn
   );
 };
@@ -257,7 +271,7 @@ export const useProposalListCount = (
   );
 
   return useQuery(
-    ["proposal_list_count", indexerGraphClient, expression],
+    [CELATONE_QUERY_KEYS.PROPOSALS_COUNT, indexerGraphClient, expression],
     queryFn
   );
 };
@@ -273,7 +287,10 @@ export const useProposalTypes = (): UseQueryResult<ProposalType[]> => {
         ),
     [indexerGraphClient]
   );
-  return useQuery(["proposal_types", indexerGraphClient], queryFn);
+  return useQuery(
+    [CELATONE_QUERY_KEYS.PROPOSAL_TYPES, indexerGraphClient],
+    queryFn
+  );
 };
 
 export interface MinDeposit {
@@ -300,7 +317,7 @@ export interface GovParams {
 export const useGovParams = (): UseQueryResult<GovParams> => {
   const lcdEndpoint = useBaseApiRoute("rest");
   const cosmwasmEndpoint = useBaseApiRoute("cosmwasm");
-  const { assetInfos } = useAssetInfos();
+  const { assetInfos } = useAssetInfos({ withPrices: false });
   const queryFn = useCallback(
     () =>
       Promise.all([
@@ -343,16 +360,20 @@ export const useGovParams = (): UseQueryResult<GovParams> => {
     [lcdEndpoint, cosmwasmEndpoint, assetInfos]
   );
 
-  return useQuery(["gov_params", lcdEndpoint, assetInfos], queryFn, {
-    keepPreviousData: true,
-    refetchOnWindowFocus: false,
-  });
+  return useQuery(
+    [CELATONE_QUERY_KEYS.GOV_PARAMS, lcdEndpoint, assetInfos],
+    queryFn,
+    {
+      keepPreviousData: true,
+      refetchOnWindowFocus: false,
+    }
+  );
 };
 
 export const useUploadAccessParams = (): UseQueryResult<UploadAccess> => {
   const cosmwasmEndpoint = useBaseApiRoute("cosmwasm");
   return useQuery(
-    ["upload_access", cosmwasmEndpoint],
+    [CELATONE_QUERY_KEYS.UPLOAD_ACCESS_PARAMS, cosmwasmEndpoint],
     () => fetchGovUploadAccessParams(cosmwasmEndpoint),
     { keepPreviousData: true, refetchOnWindowFocus: false }
   );